% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRGA_data_fxns.R
\name{combine_all}
\alias{combine_all}
\title{Combine all and format}
\usage{
combine_all(samples, interpolate_data, meta_data)
}
\arguments{
\item{interpolate_data}{This is the standard}
}
\description{
Combine all and format
}
