% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRGA_data_fxns.R
\name{import_meta}
\alias{import_meta}
\title{Import metadata}
\usage{
import_meta(fileloc)
}
\arguments{
\item{fileloc}{my file loc}
}
\description{
Import meta, data, interpolate standards and organize
}
\seealso{
Other import: \code{\link{import_sample}}
}
\concept{import}
