% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRGA_data_fxns.R
\name{import_msrmts}
\alias{import_msrmts}
\title{Import sample data}
\usage{
import_msrmts(fileloc, meta_data)
}
\arguments{
\item{meta_data}{this is the date and time metadata for this sampling}
}
\description{
This function imports all the \strong{cleaned} sample data and
separates it from standards
}
\concept{import}
