% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRGA_data_fxns.R
\name{import_sample}
\alias{import_sample}
\title{Import sample data}
\usage{
import_sample(fileloc, meta_data)
}
\arguments{
\item{fileloc}{my file loc}

\item{meta_data}{this is the date and time metadata for this sampling}
}
\description{
This function imports all the \emph{cleaned} sample data
}
\seealso{
Other import: \code{\link{import_meta}}
}
\concept{import}
