% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRGA_data_fxns.R
\name{interpolate_std}
\alias{interpolate_std}
\title{Import and interpolate standards}
\usage{
interpolate_std(samples)
}
\arguments{
\item{samples}{samples should be "samp" from the \code{\link{import_sample}} function}
}
\description{
Import and interpolate standards
}
